% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_jsonb}
\alias{as_jsonb}
\title{Convert a vector to JSONB}
\usage{
as_jsonb(x, ...)
}
\arguments{
\item{x}{the vector to convert}

\item{...}{arguments passed to yyjsonr::write_json_raw}
}
\description{
Convert a vector to JSONB
}
