% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Category}
\alias{Category}
\title{Category S7 Class}
\usage{
Category(
  id = "01KETG6YAY2FHXEK8Q5JZJHJYV",
  title = character(0),
  description = character(0)
)
}
\arguments{
\item{id}{ULID unique identifier for the category.}

\item{title}{Character title of the category.}

\item{description}{Character description of the category.}
}
\description{
An S7 class representing a category for organizing todos.
}
