% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chief-package.R
\docType{package}
\name{chief-package}
\alias{chief}
\alias{chief-package}
\title{chief: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
