% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_jsonb}
\alias{from_jsonb}
\title{Convert JSONB to an R object}
\usage{
from_jsonb(x, ...)
}
\arguments{
\item{x}{a \code{blob} object to parse}

\item{...}{arguments passed to yyjsonr::read_json_raw()}
}
\description{
Convert JSONB to an R object
}
