% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{new_category}
\alias{new_category}
\title{Create a new Category object}
\usage{
new_category(title, description = NA_character_)
}
\arguments{
\item{title}{The title of the category (required).}

\item{description}{Optional description of the category.}
}
\value{
A new Category object.
}
\description{
Create a new Category object
}
