% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{Todo}
\alias{Todo}
\title{Todo S7 Class}
\usage{
Todo(
  id = "01KETGSZVZQYSQ2ZVET4E1ZV71",
  created = (function (.data = double(), tz = "") 
 {
     .POSIXct(.data, tz = tz)
 })(),
  completed = (function (.data = double(), tz = "") 
 {
     .POSIXct(.data, tz = tz)

    })(),
  deadline = (function (.data = double(), tz = "") 
 {
     .POSIXct(.data, tz = tz)

    })(),
  priority = integer(0),
  title = character(0),
  description = character(0),
  tags = character(0),
  category = "01KETGSZVZQYSQ2ZVET4E1ZV71"
)
}
\arguments{
\item{id}{ULID unique identifier for the todo.}

\item{created}{POSIXct timestamp when the todo was created.}

\item{completed}{POSIXct timestamp when the todo was completed (NA if incomplete).}

\item{deadline}{POSIXct deadline for the todo (NA if no deadline).}

\item{priority}{Integer priority level from 1 (lowest) to 5 (highest).}

\item{title}{Character title of the todo.}

\item{description}{Character description of the todo.}

\item{tags}{Character vector of tags for categorization.}

\item{category}{ULID of the category this todo belongs to.}
}
\description{
An S7 class representing a todo item with metadata including creation time,
completion status, deadline, priority, title, description, tags, and category.
}
