% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{Ulid}
\alias{Ulid}
\title{ULID Property}
\format{
An object of class \code{S7_property} of length 6.
}
\usage{
Ulid
}
\description{
An S7 property for Universally Unique Lexicographically Sortable Identifiers (ULIDs).
ULIDs are 26-character strings that are sortable by timestamp.
}
\keyword{datasets}
